requirements.txt
fastapi
uvicorn
streamlit
transformers
torch
sqlalchemy
pandas
pydantic-settings
plotly
requests
pytest
python-json-logger
feedparser
